requirements.txt
openai==1.64.0
python-dotenv==1.0.1
ollama==0.4.7
gradio==5.19.0
langchain==0.3.19
langchain-community==0.3.18
langchain_deepseek==0.1.2
sentence-transformers==3.4.1
langchain-huggingface==0.1.2
langchain-ollama==0.2.3
faiss-cpu==1.10.0
pdfplumber==0.11.5
